Performance/Self Protection for Aura Kingdom / Twin Saga
On Aura Kingdom or Twin Saga u can increase ur Performance when u use the File Proctection on the "error.log" File.
They have any Performance problems when u doenst use that. Thats not only the reason, on the error.log File they have any Information about ur Client, if u are using on the Official Server the Injector from Alain it doenst send any Information.

You are can use this File, overrite on ur Game Path the error.log File. Make u sure the Write Proctection is activated.
or Erase all the Information on the error.log file with ur Editor, save it, activate the Write Proctection.
(It works for the Private Server to)

Cheers~